REORGANIZACION
Reorganizo un poco el proyecto, carpeta de scripts y tests dejando el script principal en la raiz del proyecto. :)
