+++ scripts_descartados/sin_numpy_exponenciacion_rapida.py
