+++ tests/test_sin_numpy_exponenciacion_rapida.py
@@ -1,9 +1,13 @@
+#test Adrian: 22004996
+
 import pytest
 import numpy as np
-from sin_numpy_exponenciacion_rapida import (
-    elevar_matriz_a_potencia_exponenciacion_rapida,
-    multiplicar_matrices,
-)
+import sys
+import os
+sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
+
+from scripts_descartados.sin_numpy_exponenciacion_rapida import elevar_matriz_a_potencia_exponenciacion_rapida, multiplicar_matrices
+
 
 
 @pytest.fixture
